+++ src/main/java/com/rinke/solutions/pinball/renderer/PcapRenderer.java
@@ -39,8 +39,8 @@ public class PcapRenderer extends Renderer {
     		Header header = pcap.readHeader();
     		System.out.println(header);
     		Paket p;
-    		long lastTimestamp = 0;
-    		long tc = 0;
+    		int lastTimestamp = 0;
+    		int tc = 0;
     		while( ( p = pcap.readPaket()) != null ) {
     			byte[] data = new byte[p.incLen];
     			stream.read(data);
